02-use-cases/code-interpreter/requirements.txt
bedrock-agentcore
boto3
fastapi
uvicorn
python-multipart
websockets
pydantic
python-dotenv
strands-agents>=0.1.8
